# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Meal Planner API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /meals/analyze-recipe:
    post:
      operationId: analyzeRecipeUrl
      tags: [meals]
      summary: Analyze a recipe URL with AI and extract ingredients
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnalyzeRecipeBody"
      responses:
        "200":
          description: Extracted ingredients with pantry availability
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyzeRecipeResult"
  /meals/save-recipe:
    post:
      operationId: saveAnalyzedRecipe
      tags: [meals]
      summary: Save an analyzed recipe as a meal and optionally assign to a weekly plan day
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveRecipeBody"
      responses:
        "201":
          description: Saved meal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Meal"
...
components:
  schemas:
...
    Ingredient:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        quantity:
          type: string
        unit:
          type: string
        category:
          type: string
          description: "produce, dairy, meat, pantry, etc."
        isCommonPantryItem:
          type: boolean
          description: Common spices/staples likely already in pantry
      required: [id, name, quantity, unit, category, isCommonPantryItem]
    Side:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
      required: [id, name, description]
...
    Meal:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        cuisine:
          type: string
        protein:
          type: string
        isGlutenFree:
          type: boolean
        cookTimeMinutes:
          type: integer
        servings:
          type: integer
        calories:
          type: integer
        imageUrl:
          type: string
          nullable: true
        isFavorited:
          type: boolean
        instructions:
          type: string
          nullable: true
        ingredients:
          type: array
          items:
            $ref: "#/components/schemas/Ingredient"
        availableSides:
          type: array
          items:
            $ref: "#/components/schemas/Side"
        tags:
          type: array
          items:
            type: string
      required:
        [
          id,
          name,
          description,
          cuisine,
          protein,
          isGlutenFree,
          cookTimeMinutes,
          servings,
          calories,
          isFavorited,
          ingredients,
          availableSides,
          tags,
        ]

    # ── WEEKLY PLAN ─────────────────────────────────────────────────────────
...
    AnalyzeRecipeBody:
      type: object
      properties:
        url:
          type: string
          description: URL of the recipe page to analyze
      required: [url]
    RecipeIngredient:
      type: object
      properties:
        name:
          type: string
        quantity:
          type: string
        unit:
          type: string
          nullable: true
        category:
          type: string
        inPantry:
          type: boolean
          description: Whether this ingredient is already in the pantry
      required: [name, quantity, category, inPantry]
    AnalyzeRecipeResult:
      type: object
      properties:
        recipeName:
          type: string
        cuisine:
          type: string
          nullable: true
        protein:
          type: string
          nullable: true
        isGlutenFree:
          type: boolean
          nullable: true
        cookTimeMinutes:
          type: integer
          nullable: true
        calories:
          type: integer
          nullable: true
        servings:
          type: integer
          nullable: true
        instructions:
          type: string
          nullable: true
          description: Full recipe instructions extracted from the page
        ingredients:
          type: array
          items:
            $ref: "#/components/schemas/RecipeIngredient"
        haveCount:
          type: integer
        needCount:
          type: integer
      required: [recipeName, ingredients, haveCount, needCount]
    SaveRecipeBody:
      type: object
      properties:
        recipeName:
          type: string
        cuisine:
          type: string
          nullable: true
        protein:
          type: string
          nullable: true
        isGlutenFree:
          type: boolean
          nullable: true
        cookTimeMinutes:
          type: integer
          nullable: true
        calories:
          type: integer
          nullable: true
        instructions:
          type: string
          nullable: true
        sourceUrl:
          type: string
          nullable: true
        assignToDay:
          type: string
          nullable: true
          description: Day of week to assign to (e.g. "monday"). Optional.
        ingredients:
          type: array
          items:
            $ref: "#/components/schemas/RecipeIngredient"
      required: [recipeName, ingredients]
...